:sparkles: Bring new mobile menu & refacto desktop & mobile
